+++ src/desktop/launcher/filesystem.js
@@ -1,7 +1,26 @@
 const { dialog } = require('electron');
-const { writeFile } = require('fs/promises');
+const { readFile, stat, writeFile } = require('fs/promises');
 
 const { getMainWindow } = require('./main-window');
+
+async function readBufferFromFile(options = {}) {
+  const { dialogOptions, maxSize = 1048576 } = options;
+  const { filePaths } = await dialog.showOpenDialog(getMainWindow(), {
+    ...dialogOptions,
+    multiSelections: false,
+  });
+
+  if (filePaths && filePaths.length === 1) {
+    const fileStat = await stat(filePaths[0]);
+    if (fileStat.size > maxSize) {
+      throw new Error(
+        `File too large; maximum allowed size is ${maxSize} bytes`
+      );
+    }
+
+    return readFile(filePaths[0]);
+  }
+}
 
 async function writeBufferToFile({ buffer, preferredFilename, dialogOptions }) {
   const { canceled, filePath } = await dialog.showSaveDialog(getMainWindow(), {
@@ -14,5 +33,6 @@ async function writeBufferToFile({ buffer, preferredFilename, dialogOptions }) {
 }
 
 module.exports = {
+  readBufferFromFile,
   writeBufferToFile,
 };